[cl][ds][input][caret] A caret position is always right-end
:Release Notes:
Made a caret position always be at right-end

:Detailed Notes:
When a user click the obscure boundary of input element, a caret can be
located at both sides according to where the user click.

In current state, if a user click the upper side boundary, a caret
is located at left-end. If the user click the bottom side boundary,
the caret is located at right-end.

QE required that a caret should be located at right-end regardless
where a user clicked.
This patch makes the caret position to be right-end always when the
boundary of input element is clicked.

:Testing Performed:
Build and test below steps.
1. launch setting > network > wired network > edit
2. click the upper boundary of input element
2-1. click the bottom boundary of input element
3. check if in the above two cases, caret position is right-end

:QA Notes:
Test cases: GT-27741

:Issues Addressed:
[TVPLAT-34980] [Migration] Correct a mispositioned caret on left end
[PLAT-117211] Port [cl][ds][input][caret] A caret position always right-end

Change-Id: I9a9bcaa634e912100b5afa2c7478eefeebef2a28
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/267120
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/283545
Tested-by: Alexander Trofimov <alexander.trofimov@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Sergey Kipet <sergey.kipet@lge.com>
Reviewed-by: Jose Dapena <jose.dapena@lge.com>
Reviewed-by: Dmitry Ketov <dmitry.ketov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>