[tv][ds][media] Add OPUS audio codec type and pass info to SMP
:Release Notes:
Add OPUS audio codec type and pass info to SMP. This Fixes video
is not playing in youtube and Shaka player

:Detailed Notes:
These OPUS changes are made in http://gpro.lge.com/269519.
But this is codec type support is not related to the other part of
the code present in same patch set. So separating them out and
adding as a new patch set. Rest of the changes from above patchset
will be done in corresponding future patch sets as needed.

:Testing Performed:
1. Launch Youtube App.
2. Play a video file. eg: https://www.youtube.com/watch?v=WbwIoQYP6no
3. Check whether it is opus audio coded using stats for nerds.
4. Video and audio should play well.
5. Launch Shaka-player App.
6. Click All Content and choose WebM in Container dropdown.
7. Play "Big Buck Bunny: the Dark Truths of a Video Dev Cartoon (DASH)"
8. Video Playback should be fine.

:QA Notes:
Test-cases: GT-27892

:Issues Addressed:
[PLAT-117535] [chromium84] Video is not playing.(Youtube,
              Shaka player)

Change-Id: I73f768957cfa56129e3badbe2581345372f5a81e
Originally-Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium79/+/269519
Reviewed-on: http://gpro.lge.com/c/webos-pro/chromium84/+/283418
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Dmitry Ketov <dmitry.ketov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Karthick Sundaram <karthick.sundaram@lge.com>
Tested-by: Test BTS <test.bts@lge.com>